Eliminated the need for cudaMemcpy by using cudaMallocManaged
cudaMallocManaged (CUDA 3 +) uses unified memory.

It may (or not) be slower than using cudaMemcpy,
but is certainly simpler to develop applications.

In particular, I think this may improve performance
in a receiver that will require signficant use of
the GPU and CPU processing the incoming signal.

I have kept the original using cudaMemcpy
in the file,  1d_convolution_tests_Memcpy.cu
for performance comparisons (rather than a git branch).